Import code from previous repository
Previous repository had contained information that needed to be
permanently removed from source files. In the interest of time,
repository was deleted and this one was created.

Application is currently able to create, retrieve, and delete records
relating to user accounts and items from the database. There are also
some tests available. Updating records is not supported at this time.

This code came off the development branch of the previous repository.
It is not yet stable or ready for release.