Remove .env from tracking and add example.env for reference
In order to eliminate the need to make changes to source when running
the server under different configurations, files related to
configurations (such as SSL keys, usernames and passwords) have been
pulled out of the source code when possible and moved to the .env file.
Going forward, .env will be in the .gitignore so that these values
can be supplied once on target machine and never have to be modified
again after pulls. Instead, example.env will provide a list of the
values that can be supplied to the manually created .env file.